Make these calendaring classes the same as in master.
I can't get the akonadi calendaring unit-tests to run on
4.11. (Is testrunner isolated tests only working on master?).

The only way I'll continue bug-fixing in 4.11 is if the tests run,
so, by making the code the same as in master, I can run be sure the
code is correct as the test pass in master.

These changes are only refactorings and such anyway.