Don't send CANCEL to attendees  when deleting an event we didn't organize.
When deleting something that's not ours, only a REPLY with
PartStat=Declined must be sent, and to the organizer only.

Due to a bug, CANCEL was being sent to everybody, as if we were
the organizer.

Unit-test will go in master.

Bug 318394 should be similar, but for modification instead of deletion.

BUG: 217211
BUG: 306755
CCBUG: 318394
FIXED-IN: 4.11.3