ITIPHandler: emit correct signal in case of an error loading the calendar.
Summary:
This block is inside if (m_currentOperation == OperationProcessiTIPMessage)
so this is the signal that the caller (SyncItipHandler in kdepim-addons) expects.

Whether one faulty calendar item should lead to a complete calendar loading
failure, a messagebox when accepting an invitation, and a failure to insert
the invitation into the calendar, is another debate (and another fix surely),
but at least the event loop in SyncItipHandler finishes on error too.

Test Plan:
Clicking on an acceptation in kmail, with (apparently) invalid
kolab calendar items in shared folders.

Reviewers: smartins

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D5107