Use the chosen organizer identity when emailing attendees
Summary:
When using multiple identities, and creating a new event with attendees,
the identity selected as organizer should be used in order to use the corresponding Sent folder.

Of course when editing an event created by someone else, this lookup will fail, and we'll
fall back to the global korganizer setting instead, as before. Hmm, maybe we should look up
which identity of ours is an attendee? Not sure about the use case though, probably only
shared calendar folders.

Test Plan:
Created an event with a non-default identity, invited someone else,
now the mail is stored in the correct Sent folder for that identity.

Reviewers: mlaurent

Reviewed By: mlaurent

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D9789