IncidenceChanger: add getter/setter for ETM
Summary:
This allows korganizer to provide an ETM, and IncidenceEditor to use it
for the Collection combobox. As a result, that combo is populated right
away when the incidence dialog shows up, rather than the user having to
wait for it to populate before being able to press Apply or OK.

Test Plan: works, together with a change in incidenceeditor and in korganizer.

Reviewers: vkrause, mlaurent, winterz

Reviewed By: mlaurent

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D28556