Allow instance creation using an item
Most `IncidenceChanger` methods take an `Item` parameter, but
`createIncidence` takes a bare incidence.  This means incidences
can't have tags when they are created.  Currently the Incidence Editor
lets the user select tags when creating a new instance, and the tags
appear in the raw payload, but they are discarded for the `createInstance`
call and so are not added to the `PimItemTagRelation` table.

This patch adds a method with an `Item` parameter to carry tags.