Don't make notifications for already ended events persistent
This affects both events that had already ended on startup as well as
active notifications while an event ended. Those are then still present
but are only listed in the notification panel after some time, putting
the focus more strongly on the upcoming/current reminders.

Reminders related to tasks are not affected by this.

# Conflicts:
#	reminder-daemon/alarmnotification.cpp