Fix crash with reminder-daemon when reminder dismissed as notifications are being iterated over
These bugs:

- https://bugs.kde.org/show_bug.cgi?id=455902
- https://bugs.kde.org/show_bug.cgi?id=456137

feature crashing when iterating over the `QHash` containing the notifications about to be emitted. This is likely caused by the fact that the user is able to dismiss a reminder when it is emitted, and dismissal removes this notification from the `QHash`. This can happen as the `QHash` is being iterated over, and boom.

This MR creates a copy of the notifications hash to iterate over when emitting notifications, preventing the crash from happening