Fix crash when opening e-mail with calendar invitation
Commit d82e7001 introduced use of new loading state API, which has
unfortunately caused a regression. The reason for it is a constructor
circular dependency. The FetchJobCalendarPrivate class constructor is
trying to call a method of a base class, which hasn't been constructed
at this point, as the CalendarBase constructor has not executed yet
waiting for the FetchJobCalendarPrivate constructor to finish.

Untangle this mess by moving the method call to the FetchJobCalendar
constructor, at which point the base class is already initialized.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>