Sign/encrypt iTIP emails
This change implements support for signing and/or encrypting iTIP
emails. The code uses the MessageComposer::Composer to compose the
signed/encrypted message, so the same code as KMail to avoid duplicating
all the complex message-building logic. There's still a lot of code
duplication between this and the MessageComposer::ComposerViewBase
class, but that's not easily avoidable without larger refactoring of the
ComposerViewBase and further separating UI & logic.

It honors the organizer's crypto preferences (which is stored in the Identity
used for sending the iTIP email) as well as crypto preferences of all
attendees (which are loaded from KAddressbook).

The overall UX of this feature is kind of incomplete - there's no way
for the user to enable singing or encryption manually like in the KMail
composer and also there's no indication in the Incidence Editor UI whether
the iTIP email will be signed or encrypted. This can be addressed separately
in the incidenceeditor later, thought and I think it doesn't prevent
this feature from being merged. Also the message viewer in KMail is not
showing the pretty iTIP preview, just shows the cal.ics file as an
attachment.

BUG: 297956