Add Akonadi::CalendarUtils::displayName overload that takes ETM instead of ETMCalendar
The ETMCalendar version was relying on ETM under the hood anyway, the
new overload allows to resolve calendar name from Collection if you only
have ETM pointer at hand, not the entire ETMCalendar.

There's an additional overload that only takes the Collection, which is
to be used when the caller doesn't have a valid pointer to ETMCalendar
or ETM, since passing nullptr as the first argument would be ambiguous,
although as far as I could find, such only call exists in the
SingleCollectionCalendar class.