KalendarAC: add option to only notify about events we organize or attend
In general users only want to be notified about events they
either organize or attend. Especially in companies where people
have access to various shared calendars, they may not want to
be notified about events that are irrelevant to them.

This will be configured via an option in KOrganizer settings.