Make the SearchCollectionHelper inactive by default
We agreed that while the feature it offers is neat, it currently
doesn't work very well (mainly due to reliability on the unreliable
indexing and search). This change adds enable/disable toggle
for the class. By default the helper is disabled, and will make
sure to delete the 'Open Invitations' and 'Declined Invitations'
collections.